requirements.txt

streamlit
pandas
PyPDF2
python-docx
